questao_3_vector_store/requirements.txt
# Questão 3 – Busca Semântica com Embeddings e Vector Store

# Geração de embeddings
sentence-transformers>=2.2.0

# Vector store
faiss-cpu>=1.7.0

# Utilitários
numpy>=1.24.0

# HTTP / Web
httpx>=0.24.0
starlette>=0.27.0

# Testes
pytest>=7.0.0
